Delete prompts module, move to output.prompts
