+++ harbor_cli/commands/api/auditlog.py
@@ -18,7 +18,7 @@ from ...style import render_cli_option, render_cli_value
 from ...output.render import render_result
 from ...state import state
 from ...utils.commands import inject_resource_options
-from ...utils.prompts import check_enumeration_options
+from ...output.prompts import check_enumeration_options
 
 # Create a command group
 app = typer.Typer(
