+++ harbor_cli/commands/api/project.py
@@ -14,6 +14,8 @@ from ...logs import logger
 from ...models import MemberRoleType
 from ...models import ProjectExtended
 from ...output.console import exit_err
+from ...output.prompts import check_enumeration_options
+from ...output.prompts import delete_prompt
 from ...output.render import render_result
 from ...state import state
 from ...utils import parse_commalist
@@ -29,8 +31,6 @@ from ...utils.commands import ARG_USERNAME_OR_ID
 from ...utils.commands import inject_help
 from ...utils.commands import inject_resource_options
 from ...utils.commands import OPTION_FORCE
-from ...utils.prompts import check_enumeration_options
-from ...utils.prompts import delete_prompt
 
 # Create a command group
 app = typer.Typer(
