+++ harbor_cli/commands/api/registry.py
@@ -12,6 +12,7 @@ from harborapi.models.models import RegistryUpdate
 from ...logs import logger
 from ...output.console import exit_err
 from ...output.console import success
+from ...output.prompts import delete_prompt
 from ...output.render import render_result
 from ...state import state
 from ...utils.args import create_updated_model
@@ -19,7 +20,6 @@ from ...utils.args import model_params_from_ctx
 from ...utils.commands import inject_help
 from ...utils.commands import inject_resource_options
 from ...utils.commands import OPTION_FORCE
-from ...utils.prompts import delete_prompt
 
 # Create a command group
 app = typer.Typer(
