+++ harbor_cli/commands/api/user.py
@@ -11,6 +11,7 @@ from harborapi.models.models import UserResp
 
 from ...exceptions import HarborCLIError
 from ...logs import logger
+from ...output.prompts import delete_prompt
 from ...output.render import render_result
 from ...state import state
 from ...utils.args import create_updated_model
@@ -22,7 +23,6 @@ from ...utils.commands import OPTION_LIMIT
 from ...utils.commands import OPTION_PAGE
 from ...utils.commands import OPTION_PAGE_SIZE
 from ...utils.commands import OPTION_QUERY
-from ...utils.prompts import delete_prompt
 
 # Create a command group
 app = typer.Typer(
