+++ harbor_cli/commands/api/usergroup.py
@@ -8,11 +8,11 @@ from harborapi.models.models import UserGroup
 from ...logs import logger
 from ...models import UserGroupType
 from ...output.console import exit_err
+from ...output.prompts import delete_prompt
 from ...output.render import render_result
 from ...state import state
 from ...utils.commands import inject_resource_options
 from ...utils.commands import OPTION_FORCE
-from ...utils.prompts import delete_prompt
 
 # Create a command group
 app = typer.Typer(
